Check syntax helper for Claude
